requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlmodel>=0.0.8
httpx>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-cov>=4.1.0
pydantic>=1.10.7
alembic>=1.10.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.1.0
psycopg2-binary>=2.9.6  # For PostgreSQL
pandas>=2.0.0  # For data manipulation and analysis
matplotlib>=3.7.1  # For chart generation on the server side
reportlab>=4.0.4  # For PDF generation
weasyprint>=59.0  # For HTML to PDF conversion
numpy>=1.24.0  # For numerical operations
scikit-learn>=1.2.2  # For machine learning models
aiosmtplib>=2.0.0  # For email functionality
fastapi-mail>=1.3.0  # For sending emails 